+++ packages/effects-core/src/utils/index.ts
@@ -86,6 +86,11 @@ export function isObject (obj: any) {
   return Object.prototype.toString.call(obj) === '[object Object]';
 }
 
+export function isCanvas (canvas: HTMLCanvasElement) {
+  // 小程序 Canvas 无法使用 instanceof HTMLCanvasElement 判断
+  return typeof canvas === 'object' && canvas !== null && canvas.tagName?.toUpperCase() === 'CANVAS';
+}
+
 export function deepClone (obj: any): any {
   if (isArray(obj)) {
     return obj.map(deepClone);
@@ -119,4 +124,39 @@ export function throwDestroyedError () {
 
 export function generateGUID (): string {
   return uuidv4().replace(/-/g, '');
+}
+
+export function base64ToFile (
+  base64: string,
+  filename = 'base64File',
+  contentType = '',
+) {
+  // 去掉 Base64 字符串的 Data URL 部分（如果存在）
+  const base64WithoutPrefix = base64.split(',')[1] || base64;
+
+  // 将 base64 编码的字符串转换为二进制字符串
+  const byteCharacters = atob(base64WithoutPrefix);
+  // 创建一个 8 位无符号整数值的数组，即“字节数组”
+  const byteArrays = [];
+
+  // 切割二进制字符串为多个片段，并将每个片段转换成一个字节数组
+  for (let offset = 0; offset < byteCharacters.length; offset += 512) {
+    const slice = byteCharacters.slice(offset, offset + 512);
+    const byteNumbers = new Array(slice.length);
+
+    for (let i = 0; i < slice.length; i++) {
+      byteNumbers[i] = slice.charCodeAt(i);
+    }
+    const byteArray = new Uint8Array(byteNumbers);
+
+    byteArrays.push(byteArray);
+  }
+
+  // 使用字节数组创建 Blob 对象
+  const blob = new Blob(byteArrays, { type: contentType });
+
+  // 创建 File 对象
+  const file = new File([blob], filename, { type: contentType });
+
+  return file;
 }
